+++ lib/ruboty/github.rb
@@ -1,13 +1,16 @@
-require "active_support/core_ext/string/strip"
-require "octokit"
+# frozen_string_literal: true
 
-require "ruboty"
-require "ruboty/github/actions/base"
-require "ruboty/github/actions/close_issue"
-require "ruboty/github/actions/create_issue"
-require "ruboty/github/actions/create_pull_request"
-require "ruboty/github/actions/create_deploy_pull_request"
-require "ruboty/github/actions/merge_pull_request"
-require "ruboty/github/actions/remember"
-require "ruboty/github/version"
-require "ruboty/handlers/github"
+require 'active_support/core_ext/string/strip'
+require 'octokit'
+
+require 'ruboty'
+require 'ruboty/github/actions/base'
+require 'ruboty/github/actions/close_issue'
+require 'ruboty/github/actions/create_issue'
+require 'ruboty/github/actions/create_pull_request'
+require 'ruboty/github/actions/create_deploy_pull_request'
+require 'ruboty/github/actions/merge_pull_request'
+require 'ruboty/github/actions/remember'
+require 'ruboty/github/actions/search_issues'
+require 'ruboty/github/version'
+require 'ruboty/handlers/github'