+++ lib/ruboty/github/actions/remember.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Ruboty
   module Github
     module Actions