+++ spec/spec_helper.rb
@@ -1,11 +1,12 @@
-require "simplecov"
+# frozen_string_literal: true
+
+require 'simplecov'
 SimpleCov.start
 
-require "ruboty/github"
-require "webmock/rspec"
+require 'ruboty/github'
+require 'webmock/rspec'
 
 RSpec.configure do |config|
-  config.treat_symbols_as_metadata_keys_with_true_values = true
   config.run_all_when_everything_filtered = true
   config.filter_run :focus
 end