+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add `search issues` command
 - Support test on GitHub Actions
 - Add rubocop, rubocop-rake, rubocop-performance, rubocop-rspec gems
+- Use rubocop autocorrect and modify the code
 
 ## 0.2.3
 - Fix typo