+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+- Add `search issues` command
+- Support test on GitHub Actions
+- Add rubocop, rubocop-rake, rubocop-performance, rubocop-rspec gems
+- Use rubocop autocorrect and modify the code
+
 ## 0.2.3
 - Fix typo