+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Support test on GitHub Actions
 - Add rubocop, rubocop-rake, rubocop-performance, rubocop-rspec gems
 - Use rubocop autocorrect and modify the code
+- Added "search issues" description and status badge to README.md
 
 ## 0.2.3
 - Fix typo